Restructured code base. New package 'flashcards' to handle all the logic. Can create a separate package for config.
