+++ flashcards/cardfinder.go
@@ -1,4 +1,4 @@
-package main
+package flashcards
 
 import (
 	"fmt"
