--- testfile.txt
@@ -1,18 +1,0 @@
-What's the date? | 19/09/2024
-How are you?? | I'm good
-What is the time? | It is '15:22' 
-what is the weather? | amazing weather
-where is waldo? | to your left
-What is the date today? | 20/08/2025
-hi | hu
-What is the time? | 11:05
-What day is it today? | Monday
-Where | are
-ss | ff
-What did you change? | I removed flags :)
-Did you update the REPL mode as well | Yes, I did.
-Are you addding another flashcard? | yes
-sdsdsd | 24
-dfjhdfkjfh | 9
-crkjfrkfj | djwkjdew
-What is larger Natural Numbers or Real numbers? | Real numbers
